Create a task.install.rpm submodule
This will hold most of the code specific to RPM-based systems
